% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toUniformat.R
\name{toUniformat}
\alias{toUniformat}
\alias{to.uniformat}
\title{Creates a Uniformat file from a dataframe}
\usage{
toUniformat(typing, fname=NULL, loci=c('A', 'B', 'C', 'DRB1', 'DQB1', 'DPB1'), dict=NULL)
}
\arguments{
\item{typing}{a dataframe with first column [IND] the identifiers and the allele information in 
paired locus columns (Locus_1, Locus_2); unknown alleles should be [NA]}

\item{fname}{an optional filename to save the output; if [NULL]the output will be returned instead}

\item{loci}{the loci to be exported in the Uniformat file}

\item{dict}{the list dictionary containing the correspondence of ambiguous alleles}
}
\value{
res Uniformat formatted data
}
\description{
Creates a .unif file containing the genetic infomation from a dataframe
}
\author{
William Lemieux \email{william.lemieux@hema-quebec.qc.ca}
}
